+++ wasmtimexec/example/main.go
@@ -33,7 +33,7 @@ func (instance *Instance) Error(format string, params ...interface{}) {
 	log.Printf("ERROR: "+format+"\n", params...)
 }
 
-func (instance *Instance) GetSP() (int32, error) {
+func (instance *Instance) GetSP() (uint32, error) {
 	val, err := instance.spFn.Call(instance.store)
 	if err != nil {
 		return 0, err
@@ -44,7 +44,7 @@ func (instance *Instance) GetSP() (int32, error) {
 		return 0, fmt.Errorf("getsp: %T: expected an int32 return value", sp)
 	}
 
-	return sp, nil
+	return uint32(sp), nil
 }
 
 func (instance *Instance) Resume() error {
